split out UI logic into its own class, raycast through multiple colliders
