+++ Basis/Packages/com.basis.framework/UI/BasisUIRaycast.cs
@@ -1,0 +1,502 @@
+using Basis.Scripts.Addressable_Driver;
+using Basis.Scripts.Addressable_Driver.Enums;
+using Basis.Scripts.Addressable_Driver.Factory;
+using Basis.Scripts.Addressable_Driver.Resource;
+using Basis.Scripts.BasisSdk.Helpers;
+using Basis.Scripts.Device_Management;
+using Basis.Scripts.Device_Management.Devices;
+using Basis.Scripts.Drivers;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using UnityEngine;
+using UnityEngine.EventSystems;
+using UnityEngine.UI;
+
+namespace Basis.Scripts.UI
+{
+    public partial class BasisUIRaycast
+    {
+        public BasisPointRaycaster BasisPointRaycaster;
+
+        public LayerMask UIMask;
+
+        public RaycastHit PhysicHit;
+        public Material lineMaterial;
+        public float lineWidth = 0.01f;
+        public LineRenderer LineRenderer;
+        public bool CachedLinerRenderState = false;
+        public bool HasLineRenderer = false;
+        public bool HasRedicalRenderer = false;
+        public static string LoadMaterialAddress = "Assets/UI/Material/RayCastMaterial.mat";
+        public static string LoadUIRedicalAddress = "Assets/UI/Prefabs/highlightQuad.prefab";
+        public GameObject highlightQuadInstance;
+        public Canvas FoundCanvas;
+        public RaycastResult RaycastResult = new RaycastResult();
+        public BasisInput BasisInput;
+        private BasisDeviceMatchSettings BasisDeviceMatchableNames;
+        public BasisPointerEventData CurrentEventData;
+        public bool HadRaycastUITarget = false;
+        public bool WasCorrectLayer = false;
+        static readonly Vector3[] s_Corners = new Vector3[4];
+        static readonly RaycastUIHitComparer s_RaycastHitComparer = new RaycastUIHitComparer();
+        [SerializeField]
+        public List<RaycastUIHitData> SortedGraphics = new List<RaycastUIHitData>();
+        [SerializeField]
+        public List<RaycastResult> SortedRays = new List<RaycastResult>();
+        public bool IgnoreReversedGraphics;
+        const string k_UILayer = "UI";
+        
+        [System.Serializable]
+        public struct RaycastUIHitData
+        {
+            public RaycastUIHitData(Graphic graphic, Vector3 worldHitPosition, Vector2 screenPosition, float distance, int displayIndex)
+            {
+                this.graphic = graphic;
+                this.worldHitPosition = worldHitPosition;
+                this.screenPosition = screenPosition;
+                this.distance = distance;
+                this.displayIndex = displayIndex;
+            }
+            [SerializeField]
+            public Graphic graphic;
+            [SerializeField]
+            public Vector3 worldHitPosition;
+            [SerializeField]
+            public Vector2 screenPosition;
+            [SerializeField]
+            public float distance;
+            [SerializeField]
+            public int displayIndex;
+        }
+        sealed class RaycastUIHitComparer : IComparer<RaycastUIHitData>
+        {
+            public int Compare(RaycastUIHitData a, RaycastUIHitData b)
+                => b.graphic.depth.CompareTo(a.graphic.depth);
+        }
+
+        public async Task Initialize(BasisInput basisInput, BasisPointRaycaster pointRaycaster)
+        {
+            CurrentEventData = new BasisPointerEventData(EventSystem.current);
+            BasisInput = basisInput;
+            BasisPointRaycaster = pointRaycaster;
+            BasisDeviceMatchableNames = BasisInput.BasisDeviceMatchableNames;
+            ApplyStaticDataToRaycastResult();
+
+            UIMask = LayerMask.NameToLayer(k_UILayer);
+
+
+            HasLineRenderer = false;
+            HasRedicalRenderer = false;
+            // Create the ray with the adjusted starting position and direction
+            if (BasisDeviceMatchableNames.HasRayCastVisual)
+            {
+                // Add a Line Renderer component to the GameObject
+                LineRenderer = BasisHelpers.GetOrAddComponent<LineRenderer>(BasisPointRaycaster.gameObject);
+                LineRenderer.enabled = false;
+                AddressableGenericResource Resource = new AddressableGenericResource(LoadMaterialAddress, AddressableExpectedResult.SingleItem);
+                if (await AddressableLoadFactory.LoadAddressableResourceAsync<Material>(Resource))
+                {
+                    lineMaterial = (Material)Resource.Handles[0].Result;
+                    // Set the Line Renderer properties
+                    LineRenderer.material = lineMaterial;
+                }
+                LineRenderer.startWidth = lineWidth;
+                LineRenderer.endWidth = lineWidth;
+
+                // Set the number of points in the Line Renderer
+                LineRenderer.positionCount = 2;
+                HasLineRenderer = true;
+                LineRenderer.enabled = HasLineRenderer;
+                LineRenderer.numCapVertices = 12;
+                LineRenderer.numCornerVertices = 12;
+                LineRenderer.gameObject.layer = LayerMask.NameToLayer(k_UILayer);
+            }
+            if (BasisDeviceMatchableNames.HasRayCastRedical)
+            {
+                await CreateRadical();
+                HasRedicalRenderer = true;
+            }
+            CachedLinerRenderState = HasLineRenderer;
+
+        }
+        public async Task CreateRadical()
+        {
+            ChecksRequired Required = new ChecksRequired
+            {
+                UseContentRemoval = false,
+                DisableAnimatorEvents = false
+            };
+            (List<GameObject>, AddressableGenericResource) data = await AddressableResourceProcess.LoadAsGameObjectsAsync(LoadUIRedicalAddress, new UnityEngine.ResourceManagement.ResourceProviders.InstantiationParameters(), Required);
+            List<GameObject> gameObjects = data.Item1;
+            if (gameObjects == null)
+            {
+                return;
+            }
+            if (gameObjects.Count != 0)
+            {
+                foreach (GameObject gameObject in gameObjects)
+                {
+                    gameObject.name = BasisDeviceMatchableNames.DeviceID + "_Redical";
+                    gameObject.transform.SetParent(BasisPointRaycaster.gameObject.transform);
+                    highlightQuadInstance = gameObject;
+                    if (highlightQuadInstance.TryGetComponent(out Canvas Canvas))
+                    {
+                        Canvas.worldCamera = BasisLocalCameraDriver.Instance.Camera;
+                    }
+                }
+            }
+        }
+        public void ApplyStaticDataToRaycastResult()
+        {
+            RaycastResult.displayIndex = 0;
+            RaycastResult.index = 0;
+            RaycastResult.depth = 0;
+            RaycastResult.module = BasisPointRaycaster;
+        }
+        public void HandleUIRaycast()
+        {
+            HadRaycastUITarget = BasisPointRaycaster.PhysicHitCount > 0;
+            
+            if (HadRaycastUITarget)
+            {
+                HandleDidHit();
+            }
+            else
+            {
+                HandleNoHit();
+            }
+        }
+        private void HandleNoHit()
+        {
+            ResetRenderers();
+            RaycastResult = new RaycastResult();
+            PhysicHit = new RaycastHit();
+        }
+        private void HandleDidHit()
+        {
+            SortedGraphics.Clear();
+            SortedRays.Clear();
+
+            // only first hit counted for UI
+            // TODO: not do this?
+            PhysicHit = BasisPointRaycaster.PhysicHits[0];
+
+            PhysicHit.transform.GetComponentsInChildren<Canvas>(false, Results);
+
+            WasCorrectLayer = PhysicHit.transform.gameObject.layer == UIMask;
+            if (WasCorrectLayer)
+            {
+                RaycastToUI();
+                UpdateRayCastResult();//sets all RaycastResult data
+                UpdateLineRenderer();//updates the line denderer
+                UpdateRadicalRenderer();// moves the redical renderer
+            }
+            else
+            {
+                ResetRenderers();
+            }
+        }
+        
+        public Vector2 ScreenPoint;
+        public bool UseWorldPosition = true;
+        private void UpdateRayCastResult()
+        {
+            RaycastResult.gameObject = PhysicHit.transform.gameObject;
+            RaycastResult.distance = PhysicHit.distance;
+            if (UseWorldPosition)
+            {
+                ScreenPoint = BasisLocalCameraDriver.Instance.Camera.WorldToScreenPoint(BasisPointRaycaster.transform.position, Camera.MonoOrStereoscopicEye.Mono);
+            }
+            else
+            {
+                // we assign screenpoint manually example in BasisLocalCameraDriver
+            }
+            RaycastResult.screenPosition = ScreenPoint;
+            FoundCanvas = PhysicHit.transform.gameObject.GetComponentInParent<Canvas>();
+            if (FoundCanvas != null)
+            {
+                RaycastResult.sortingLayer = FoundCanvas.sortingLayerID;
+                RaycastResult.sortingOrder = FoundCanvas.sortingOrder;
+            }
+            RaycastResult.worldPosition = BasisPointRaycaster.ray.origin + BasisPointRaycaster.ray.direction * PhysicHit.distance;
+            RaycastResult.worldNormal = PhysicHit.normal;
+        }
+        private void UpdateLineRenderer()
+        {
+            if (HasLineRenderer && !CachedLinerRenderState)
+            {
+                LineRenderer.enabled = true;
+                CachedLinerRenderState = true;
+            }
+            else if (!HasLineRenderer && CachedLinerRenderState)
+            {
+                LineRenderer.enabled = false;
+                CachedLinerRenderState = false;
+            }
+
+            if (HasLineRenderer)
+            {
+                LineRenderer.SetPosition(0, BasisPointRaycaster.ray.origin);
+                LineRenderer.SetPosition(1, PhysicHit.point);
+            }
+        }
+
+        private void UpdateRadicalRenderer()
+        {
+            if (HasRedicalRenderer)
+            {
+                if (PhysicHit.transform != null)
+                {
+                    if (BasisDeviceManagement.IsUserInDesktop() && BasisCursorManagement.ActiveLockState() != CursorLockMode.Locked)
+                    {
+                        highlightQuadInstance.SetActive(false);
+                    }
+                    else
+                    {
+                        highlightQuadInstance.SetActive(true);
+                        highlightQuadInstance.transform.SetPositionAndRotation(PhysicHit.point, Quaternion.LookRotation(PhysicHit.normal));
+                    }
+                }
+                else
+                {
+                    highlightQuadInstance.SetActive(false);
+                }
+            }
+        }
+
+        private void ResetRenderers()
+        {
+            if (CachedLinerRenderState && HasLineRenderer)
+            {
+                LineRenderer.enabled = false;
+                CachedLinerRenderState = false;
+            }
+
+            if (HasRedicalRenderer)
+            {
+                highlightQuadInstance.SetActive(false);
+            }
+        }
+        
+        public Vector3 LastRotation;
+        public List<Canvas> Results = new List<Canvas>();
+        
+        public bool RaycastToUI()
+        {
+            Results.Sort((c1, c2) => c2.sortingOrder.CompareTo(c1.sortingOrder));
+            int Count = Results.Count;
+            for (int Index = 0; Index < Count; Index++)
+            {
+                Canvas CurrentTopLevel = Results[Index];
+                if (CurrentTopLevel != null)
+                {
+                    if(CurrentTopLevel.worldCamera == null)
+                    {
+                        CurrentTopLevel.worldCamera = BasisLocalCameraDriver.Instance.Camera;
+                    }
+                    SortedRaycastGraphics(CurrentTopLevel, CurrentTopLevel.worldCamera, ref SortedGraphics);
+                    ProcessSortedHitsResults(CurrentTopLevel, true, SortedGraphics, SortedRays);
+                    if (SortedGraphics.Count != 0)
+                    {
+                        return true;
+                    }
+                }
+            }
+            return false;
+        }
+        public void Sort<T>(IList<T> hits, IComparer<T> comparer) where T : struct => Sort(hits, comparer, hits.Count);
+        public bool ProcessSortedHitsResults(Canvas canvas, bool hitSomething, List<RaycastUIHitData> raycastHitDatums, List<RaycastResult> resultAppendList)
+        {
+            // Now that we have a list of sorted hits, process any extra settings and filters.
+            foreach (var hitData in raycastHitDatums)
+            {
+                var validHit = true;
+
+                if (hitData.graphic == null)
+                {
+                    continue;
+                }
+                var go = hitData.graphic.gameObject;
+                if (IgnoreReversedGraphics)
+                {
+                    var forward = BasisPointRaycaster.ray.direction;
+                    var goDirection = go.transform.rotation * Vector3.forward;
+                    validHit = Vector3.Dot(forward, goDirection) > 0;
+                }
+
+                validHit &= hitData.distance < BasisPointRaycaster.MaxDistance;
+
+                if (validHit)
+                {
+                    var trans = go.transform;
+                    var transForward = trans.forward;
+                    var castResult = new RaycastResult
+                    {
+                        gameObject = go,
+                        module = BasisPointRaycaster,
+                        distance = hitData.distance,
+                        index = resultAppendList.Count,
+                        depth = hitData.graphic.depth,
+                        sortingLayer = canvas.sortingLayerID,
+                        sortingOrder = canvas.sortingOrder,
+                        worldPosition = hitData.worldHitPosition,
+                        worldNormal = -transForward,
+                        screenPosition = hitData.screenPosition,
+                        displayIndex = hitData.displayIndex,
+                    };
+                    resultAppendList.Add(castResult);
+
+                    hitSomething = true;
+                }
+            }
+
+            return hitSomething;
+        }
+        public static void Sort<T>(IList<T> hits, IComparer<T> comparer, int count) where T : struct
+        {
+            if (count <= 1)
+                return;
+
+            bool fullPass;
+            do
+            {
+                fullPass = true;
+                for (var i = 1; i < count; ++i)
+                {
+                    var result = comparer.Compare(hits[i - 1], hits[i]);
+                    if (result > 0)
+                    {
+                        (hits[i - 1], hits[i]) = (hits[i], hits[i - 1]);
+                        fullPass = false;
+                    }
+                }
+            } while (fullPass == false);
+        }
+        public void SortedRaycastGraphics(Canvas canvas, Camera eventCamera, ref List<RaycastUIHitData> results)
+        {
+            var graphics = GraphicRegistry.GetGraphicsForCanvas(canvas);
+
+            results.Clear();
+            for (int i = 0; i < graphics.Count; ++i)
+            {
+                var graphic = graphics[i];
+
+                if (!ShouldTestGraphic(graphic, BasisPointRaycaster.Mask))
+                    continue;
+
+                var raycastPadding = graphic.raycastPadding;
+
+                if (RayIntersectsRectTransform(graphic.rectTransform, raycastPadding, BasisPointRaycaster.ray, out var worldPos, out var distance))
+                {
+                    if (distance <= BasisPointRaycaster.MaxDistance)
+                    {
+                        Vector2 screenPos = eventCamera.WorldToScreenPoint(worldPos);
+                        // mask/image intersection - See Unity docs on eventAlphaThreshold for when this does anything
+                        if (graphic.Raycast(screenPos, eventCamera))
+                        {
+                            results.Add(new RaycastUIHitData(graphic, worldPos, screenPos, distance, eventCamera.targetDisplay));
+                        }
+                    }
+                }
+            }
+
+            Sort(results, s_RaycastHitComparer);
+        }
+
+        public bool ShouldTestGraphic(Graphic graphic, LayerMask layerMask)
+        {
+            // -1 means it hasn't been processed by the canvas, which means it isn't actually drawn
+            if (graphic.depth == -1 || !graphic.raycastTarget || graphic.canvasRenderer.cull)
+                return false;
+
+            if (((1 << graphic.gameObject.layer) & layerMask) == 0)
+                return false;
+
+            return true;
+        }
+
+        public bool SphereIntersectsRectTransform(RectTransform transform, Vector4 raycastPadding, Vector3 from, out Vector3 worldPosition, out float distance)
+        {
+            var plane = GetRectTransformPlane(transform, raycastPadding, s_Corners);
+            var closestPoint = plane.ClosestPointOnPlane(from);
+            var ray = new Ray(from, closestPoint - from);
+            return RayIntersectsRectTransform(ray, plane, out worldPosition, out distance);
+        }
+
+        public bool RayIntersectsRectTransform(RectTransform transform, Vector4 raycastPadding, Ray ray, out Vector3 worldPosition, out float distance)
+        {
+            var plane = GetRectTransformPlane(transform, raycastPadding, s_Corners);
+            return RayIntersectsRectTransform(ray, plane, out worldPosition, out distance);
+        }
+
+        public bool RayIntersectsRectTransform(Ray ray, Plane plane, out Vector3 worldPosition, out float distance)
+        {
+            if (plane.Raycast(ray, out var enter))
+            {
+                var intersection = ray.GetPoint(enter);
+
+                var bottomEdge = s_Corners[3] - s_Corners[0];
+                var leftEdge = s_Corners[1] - s_Corners[0];
+                var bottomDot = Vector3.Dot(intersection - s_Corners[0], bottomEdge);
+                var leftDot = Vector3.Dot(intersection - s_Corners[0], leftEdge);
+
+                // If the intersection is right of the left edge and above the bottom edge.
+                if (leftDot >= 0f && bottomDot >= 0f)
+                {
+                    var topEdge = s_Corners[1] - s_Corners[2];
+                    var rightEdge = s_Corners[3] - s_Corners[2];
+                    var topDot = Vector3.Dot(intersection - s_Corners[2], topEdge);
+                    var rightDot = Vector3.Dot(intersection - s_Corners[2], rightEdge);
+
+                    // If the intersection is left of the right edge, and below the top edge
+                    if (topDot >= 0f && rightDot >= 0f)
+                    {
+                        worldPosition = intersection;
+                        distance = enter;
+                        return true;
+                    }
+                }
+            }
+
+            worldPosition = Vector3.zero;
+            distance = 0f;
+            return false;
+        }
+
+        public Plane GetRectTransformPlane(RectTransform transform, Vector4 raycastPadding, Vector3[] fourCornersArray)
+        {
+            GetRectTransformWorldCorners(transform, raycastPadding, fourCornersArray);
+            return new Plane(fourCornersArray[0], fourCornersArray[1], fourCornersArray[2]);
+        }
+
+        // This method is similar to RecTransform.GetWorldCorners, but with support for the raycastPadding offset.
+        public void GetRectTransformWorldCorners(RectTransform transform, Vector4 offset, Vector3[] fourCornersArray)
+        {
+            if (fourCornersArray == null || fourCornersArray.Length < 4)
+            {
+                BasisDebug.LogError("Calling GetRectTransformWorldCorners with an array that is null or has less than 4 elements.");
+                return;
+            }
+
+            // GraphicRaycaster.Raycast uses RectTransformUtility.RectangleContainsScreenPoint instead,
+            // which redirects to PointInRectangle defined in RectTransformUtil.cpp. However, that method
+            // uses the Camera to convert from the given screen point to a ray, but this class uses
+            // the ray from the Ray Interactor that feeds the event data.
+            // Offset calculation for raycastPadding from PointInRectangle method, which replaces RectTransform.GetLocalCorners.
+            var rect = transform.rect;
+            var x0 = rect.x + offset.x;
+            var y0 = rect.y + offset.y;
+            var x1 = rect.xMax - offset.z;
+            var y1 = rect.yMax - offset.w;
+            fourCornersArray[0] = new Vector3(x0, y0, 0f);
+            fourCornersArray[1] = new Vector3(x0, y1, 0f);
+            fourCornersArray[2] = new Vector3(x1, y1, 0f);
+            fourCornersArray[3] = new Vector3(x1, y0, 0f);
+
+            // Transform the local corners to world space, which is from RectTransform.GetWorldCorners.
+            var localToWorldMatrix = transform.localToWorldMatrix;
+            for (var index = 0; index < 4; ++index)
+                fourCornersArray[index] = localToWorldMatrix.MultiplyPoint(fourCornersArray[index]);
+        }
+    }
+}
