+++ Basis/Packages/com.basis.framework/UI/BasisUIRaycastProcess.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using UnityEngine;
 using UnityEngine.EventSystems;
 using static Basis.Scripts.UI.BasisPointRaycaster;
+using static Basis.Scripts.UI.BasisUIRaycast;
 
 namespace Basis.Scripts.UI
 {
@@ -46,17 +47,17 @@ public class BasisUIRaycastProcess
             for (int Index = 0; Index < DevicesCount; Index++)
             {
                 BasisInput input = Inputs[Index];
-                if (input.HasUIInputSupport && input.BasisPointRaycaster != null && input.BasisPointRaycaster.WasCorrectLayer)
+                if (input.HasRaycastSupport && input.BasisUIRaycast != null && input.BasisUIRaycast.WasCorrectLayer)
                 {
-                    EffectiveMouseAction |= input.BasisPointRaycaster.CurrentEventData.WasLastDown == false && input.InputState.Trigger == 1;
-                    if (input.BasisPointRaycaster.HadRaycastUITarget)
+                    EffectiveMouseAction |= input.BasisUIRaycast.CurrentEventData.WasLastDown == false && input.InputState.Trigger == 1;
+                    if (input.BasisUIRaycast.HadRaycastUITarget)
                     {
-                        List<RaycastHitData> hitData = input.BasisPointRaycaster.SortedGraphics;
-                        List<RaycastResult> RaycastResults = input.BasisPointRaycaster.SortedRays;
+                        List<RaycastUIHitData> hitData = input.BasisUIRaycast.SortedGraphics;
+                        List<RaycastResult> RaycastResults = input.BasisUIRaycast.SortedRays;
                         hitData.Sort((g1, g2) => g2.graphic.depth.CompareTo(g1.graphic.depth));
                         RaycastResult hit = RaycastResults[0];
                         hit.gameObject = hitData[0].graphic.gameObject;
-                        SimulateOnCanvas(hit, hitData[0], input.BasisPointRaycaster.CurrentEventData, input.InputState, input.LastState);
+                        SimulateOnCanvas(hit, hitData[0], input.BasisUIRaycast.CurrentEventData, input.InputState, input.LastState);
                         HasTarget = true;
                     }
                 }
@@ -69,13 +70,13 @@ public class BasisUIRaycastProcess
             for (int Index = 0; Index < DevicesCount; Index++)
             {
                 BasisInput input = Inputs[Index];
-                if (input.HasUIInputSupport && input.BasisPointRaycaster != null && input.BasisPointRaycaster.WasCorrectLayer)
+                if (input.HasRaycastSupport && input.BasisUIRaycast != null && input.BasisUIRaycast.WasCorrectLayer)
                 {
-                    SendUpdateEventToSelectedObject(input.BasisPointRaycaster.CurrentEventData); //needed if you want to use the keyboard
+                    SendUpdateEventToSelectedObject(input.BasisUIRaycast.CurrentEventData); //needed if you want to use the keyboard
                 }
             }
         }
-    public void SimulateOnCanvas(RaycastResult raycastResult, RaycastHitData hit, BasisPointerEventData currentEventData, BasisInputState Current, BasisInputState LastCurrent)
+    public void SimulateOnCanvas(RaycastResult raycastResult, RaycastUIHitData hit, BasisPointerEventData currentEventData, BasisInputState Current, BasisInputState LastCurrent)
     {
         if (hit.graphic != null)
         {
@@ -112,7 +113,7 @@ public class BasisUIRaycastProcess
             ProcessPointerButtonDrag(currentEventData);
         }
     }
-    public void CheckOrApplySelectedGameobject(RaycastHitData hit, BasisPointerEventData CurrentEventData)
+    public void CheckOrApplySelectedGameobject(RaycastUIHitData hit, BasisPointerEventData CurrentEventData)
     {
         if (hit.graphic != null)
         {
@@ -126,7 +127,7 @@ public class BasisUIRaycastProcess
             EventSystem.current.SetSelectedGameObject(null, CurrentEventData);
         }
     }
-    public void EffectiveMouseDown(RaycastHitData hit, BasisPointerEventData CurrentEventData)
+    public void EffectiveMouseDown(RaycastUIHitData hit, BasisPointerEventData CurrentEventData)
     {
         CurrentEventData.eligibleForClick = true;
         CurrentEventData.delta = Vector2.zero;
@@ -168,7 +169,7 @@ public class BasisUIRaycastProcess
             ExecuteEvents.Execute(dragObject, CurrentEventData, ExecuteEvents.initializePotentialDrag);
         }
     }
-    public void EffectiveMouseUp(RaycastHitData hit, BasisPointerEventData CurrentEventData)
+    public void EffectiveMouseUp(RaycastUIHitData hit, BasisPointerEventData CurrentEventData)
     {
         var target = CurrentEventData.pointerPress;
         ExecuteEvents.Execute(target, CurrentEventData, ExecuteEvents.pointerUpHandler);
