--- Basis/Packages/com.basis.framework/UI/RaycastHitComparer.cs
@@ -1,14 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Basis.Scripts.UI
-{
-public partial class BasisPointRaycaster
-{
-    /// </summary>
-    sealed class RaycastHitComparer : IComparer<RaycastHitData>
-    {
-        public int Compare(RaycastHitData a, RaycastHitData b)
-            => b.graphic.depth.CompareTo(a.graphic.depth);
-    }
-}
-}
