--- Basis/Packages/com.basis.framework/UI/RaycastHitComparer.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 483907eac9b56c84e9cffa4849197bbf
