--- Basis/Packages/com.basis.framework/UI/RaycastHitData.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f8af287b25d5ef647bacb5b5734cdf62
